started work on implementing 64 bit OS
moved crti and ctrn to be arch specific
changed PDE and PTE flags to be defines instead of enum
changed string.h to not throw a warning
added flags for 64 bit paging to kmalloc_def
enabled 64 bit in boot.s

+++ usr/include/string.h
@@ -113,7 +113,7 @@ extern "C"
       s1++;
       s2++;
     }
-    return (int)((const unsigned char *)s1 - *(const unsigned char *)s2);
+    return (ptrdiff_t)((const unsigned char *)s1 - *(const unsigned char *)s2);
   }
   static int strncmp(const char *s1, const char *s2, size_t n)
   {
@@ -122,7 +122,7 @@ extern "C"
       if (*s1 != *s2)
         break;
     }
-    return (int)((const unsigned char *)s1 - *(const unsigned char *)s2);
+    return (ptrdiff_t)((const unsigned char *)s1 - *(const unsigned char *)s2);
   }
   static char *strchr(char *s, char c)
   {
